Factor out TypeSystemThrowingILEmitter
